requirements.txt
azure-core==1.35.0
azure-identity==1.23.1
azure-storage-blob==12.26.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
idna==3.10
isodate==0.7.2
msal==1.33.0
msal-extensions==1.3.1
pycparser==2.22
PyJWT==2.10.1
requests==2.32.4
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
